added redux and refactored code
