+++ src/reducers/index.js
@@ -1,0 +1,9 @@
+import { combineReducers } from "redux";
+
+import infoReducer from "./infoReducer";
+
+const reducer = combineReducers({
+  info: infoReducer,
+});
+
+export default reducer;
