+++ src/services/api.js
@@ -1,0 +1,4 @@
+export const getAllExperience = "http://localhost:1000/experience";
+export const getAllEducation = "http://localhost:1000/education";
+export const getAllProjects = "http://localhost:1000/projects";
+export const getAllSkills = "http://localhost:1000/skills";
